sunday1_ai_browser_agent/config.py
from dotenv import load_dotenv
import os

load_dotenv()

class Config:
    OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
    OPENAI_MODEL = "gpt-3.5-turbo"
    BROWSER_HEADLESS = False
    ACTION_DELAY = 1  # seconds