sunday1_ai_browser_agent/requirements.txt
fastapi
uvicorn
openai
python-dotenv
playwright
rich
